LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/fetch/sec-metadata/object.tentative.https.sub-expected.txt



FAIL Same-Origin object assert_not_equals: Empty Sec-Metadata header. got disallowed value ""
FAIL Same-Site object assert_equals: destination expected (string) "object" but got (undefined) undefined
FAIL Cross-Site object assert_equals: destination expected (string) "object" but got (undefined) undefined
